# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Simple FastAPI ToDo List API v1",
    "description": "\n## Login Instructions\n2. **Request OTP Code**: To log in, request an OTP code by posting to `login/code/` with your phone number.\n3. **Submit Phone and Code**: Finally, post your phone number and the received OTP code to complete the login process.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/todos/invites/": {
      "get": {
        "tags": [
          "todos",
          "invites"
        ],
        "summary": "Read Invites",
        "description": "Retrieve invites.",
        "operationId": "todos-read_invites",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todos",
          "invites"
        ],
        "summary": "Create Invite",
        "description": "Create new invite.",
        "operationId": "todos-create_invite",
        "parameters": [
          {
            "name": "todo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Todo Id"
            }
          },
          {
            "name": "user_phone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/todos/invites/{invite_id}/accept": {
      "post": {
        "tags": [
          "todos",
          "invites"
        ],
        "summary": "Accept Invite",
        "description": "Accept invite by ID.",
        "operationId": "todos-accept_invite",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/todos/invites/{invite_id}/decline": {
      "post": {
        "tags": [
          "todos",
          "invites"
        ],
        "summary": "Decline Invite",
        "description": "Accept invite by ID.",
        "operationId": "todos-decline_invite",
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/todos/": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Read Todos",
        "description": "Retrieve todos.",
        "operationId": "todos-read_todos",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodosOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todos"
        ],
        "summary": "Create Todo",
        "description": "Create new todo.",
        "operationId": "todos-create_todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/todos/{todo_id}": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Read Todo",
        "description": "Get todo by ID.",
        "operationId": "todos-read_todo",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "todos"
        ],
        "summary": "Update Todo",
        "description": "Update a todo.",
        "operationId": "todos-update_todo",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Todo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "todos"
        ],
        "summary": "Delete Todo",
        "description": "Delete a todo.",
        "operationId": "todos-delete_todo",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageOutShort": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "blur_hash": {
            "type": "string",
            "title": "Blur Hash"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "blur_hash",
          "url"
        ],
        "title": "ImageOutShort"
      },
      "InviteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "todo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TodoOutShort"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "todo"
        ],
        "title": "InviteOut"
      },
      "InvitesOut": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/InviteOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "count",
          "total",
          "limit",
          "skip",
          "data"
        ],
        "title": "InvitesOut"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
      "TodoCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus",
            "default": "new"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoCreate"
      },
      "TodoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Users",
            "default": []
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "modified_at"
        ],
        "title": "TodoOut"
      },
      "TodoOutShort": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status"
        ],
        "title": "TodoOutShort"
      },
      "TodoStatus": {
        "type": "string",
        "enum": [
          "new",
          "in_progress",
          "done"
        ],
        "title": "TodoStatus"
      },
      "TodoUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus",
            "default": "new"
          },
          "user_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Ids"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoUpdate"
      },
      "TodosOut": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TodoOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "count",
          "total",
          "limit",
          "skip",
          "data"
        ],
        "title": "TodosOut"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "phone": {
            "type": "integer",
            "title": "Phone"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "profile_image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageOutShort"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone"
        ],
        "title": "UserOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}